+++ Common/CommonHelper.cs
@@ -32,6 +32,9 @@ namespace Pathoschild.Stardew.Common
             return pixel;
         });
 
+        /// <summary>The width of the borders drawn by <see cref="DrawTab"/>.</summary>
+        public const int ButtonBorderWidth = 4 * Game1.pixelZoom;
+
 
         /*********
         ** Accessors
@@ -179,6 +182,39 @@ namespace Pathoschild.Stardew.Common
             spriteBatch.DrawTextBlock(Game1.smallFont, label, position + new Vector2(gutterSize), wrapWidth); // draw again over texture box
 
             return labelSize + new Vector2(paddingSize);
+        }
+
+        /// <summary>Draw a tab texture to the screen.</summary>
+        /// <param name="spriteBatch">The sprite batch to which to draw.</param>
+        /// <param name="x">The X position at which to draw.</param>
+        /// <param name="y">The Y position at which to draw.</param>
+        /// <param name="innerWidth">The width of the button's inner content.</param>
+        /// <param name="innerHeight">The height of the button's inner content.</param>
+        /// <param name="innerDrawPosition">The position at which the content should be drawn.</param>
+        /// <param name="align">The button's horizontal alignment relative to <paramref name="x"/>. The possible values are 0 (left), 1 (center), or 2 (right).</param>
+        /// <param name="alpha">The button opacity, as a value from 0 (transparent) to 1 (opaque).</param>
+        /// <param name="forIcon">Whether the button will contain an icon instead of text.</param>
+        public static void DrawTab(SpriteBatch spriteBatch, int x, int y, int innerWidth, int innerHeight, out Vector2 innerDrawPosition, int align = 0, float alpha = 1, bool forIcon = false)
+        {
+            // calculate outer coordinates
+            int outerWidth = innerWidth + CommonHelper.ButtonBorderWidth * 2;
+            int outerHeight = innerHeight + Game1.tileSize / 3;
+            int offsetX = align switch
+            {
+                1 => -outerWidth / 2,
+                2 => -outerWidth,
+                _ => 0
+            };
+
+            // calculate inner coordinates
+            {
+                int iconOffsetX = forIcon ? -Game1.pixelZoom : 0;
+                int iconOffsetY = forIcon ? 2 * -Game1.pixelZoom : 0;
+                innerDrawPosition = new Vector2(x + CommonHelper.ButtonBorderWidth + offsetX + iconOffsetX, y + CommonHelper.ButtonBorderWidth + iconOffsetY);
+            }
+
+            // draw texture
+            IClickableMenu.drawTextureBox(spriteBatch, Game1.menuTexture, new Rectangle(0, 256, 60, 60), x + offsetX, y, outerWidth, outerHeight + Game1.tileSize / 16, Color.White * alpha);
         }
 
         /// <summary>Draw a button background.</summary>