+++ Common/UI/DropdownList.cs
@@ -24,14 +24,14 @@ namespace Pathoschild.Stardew.Common.UI
         /****
         ** Items
         ****/
-        /// <summary>The index of the selected item.</summary>
-        private int SelectedIndex;
-
-        /// <summary>The selected item.</summary>
-        private TItem SelectedItem => this.Items.First(p => p.Index == this.SelectedIndex).Value;
+        /// <summary>The selected entry.</summary>
+        private DropListItem Selected;
 
         /// <summary>The items in the list.</summary>
         private readonly DropListItem[] Items;
+
+        /// <summary>The clickable components representing the list items.</summary>
+        private readonly List<ClickableComponent> ItemComponents = new List<ClickableComponent>();
 
         /// <summary>The item index shown at the top of the list.</summary>
         private int FirstVisibleIndex;
@@ -42,20 +42,34 @@ namespace Pathoschild.Stardew.Common.UI
         /// <summary>The maximum index for the first item.</summary>
         private int MaxFirstVisibleIndex;
 
+        /// <summary>Get the display name for a value.</summary>
+        private readonly Func<TItem, string> GetLabel;
+
+
         /****
         ** Rendering
         ****/
-        /// <summary>The dropdown's origin position.</summary>
-        private readonly Vector2 Origin;
-
         /// <summary>The font with which to render text.</summary>
         private readonly SpriteFont Font;
 
-        /// <summary>The height of text rendered in the <see cref="Font"/>.</summary>
-        private readonly int FontHeight;
 
-        /// <summary>The clickable components representing the list items.</summary>
-        private readonly List<ClickableComponent> ItemComponents = new List<ClickableComponent>();
+        /*********
+        ** Accessors
+        *********/
+        /// <summary>The selected item.</summary>
+        public TItem SelectedItem => this.Selected.Value;
+
+        /// <summary>The display label for the selected item.</summary>
+        public string SelectedLabel => this.GetLabel(this.SelectedItem);
+
+        /// <summary>The maximum height for the possible labels.</summary>
+        public int MaxLabelHeight { get; }
+
+        /// <summary>The maximum width for the possible labels.</summary>
+        public int MaxLabelWidth { get; private set; }
+
+        /// <summary>The <see cref="ClickableComponent.myID"/> value for the top entry in the dropdown.</summary>
+        public int TopComponentId => this.ItemComponents[0].myID;
 
 
         /*********
@@ -64,23 +78,30 @@ namespace Pathoschild.Stardew.Common.UI
         /// <summary>Construct an instance.</summary>
         /// <param name="selectedItem">The selected item.</param>
         /// <param name="items">The items in the list.</param>
-        /// <param name="nameSelector">A lambda which returns the display name for an item.</param>
+        /// <param name="getLabel">Get the display label for an item.</param>
         /// <param name="x">The X-position from which to render the list.</param>
         /// <param name="y">The Y-position from which to render the list.</param>
         /// <param name="font">The font with which to render text.</param>
-        public DropdownList(TItem selectedItem, TItem[] items, Func<TItem, string> nameSelector, int x, int y, SpriteFont font)
+        public DropdownList(TItem selectedItem, TItem[] items, Func<TItem, string> getLabel, int x, int y, SpriteFont font)
             : base(new Rectangle(), nameof(DropdownList<TItem>))
         {
             // save values
-            this.SelectedIndex = Array.IndexOf(items, selectedItem);
             this.Items = items
-                .Select((item, index) => new DropListItem(index, nameSelector(item), item))
+                .Select((item, index) => new DropListItem(index, getLabel(item), item))
                 .ToArray();
             this.Font = font;
-            this.FontHeight = (int)font.MeasureString("abcdefghijklmnopqrstuvwxyz").Y;
+            this.MaxLabelHeight = (int)font.MeasureString("abcdefghijklmnopqrstuvwxyz").Y;
+            this.GetLabel = getLabel;
+
+            // set initial selection
+            int selectedIndex = Array.IndexOf(items, selectedItem);
+            this.Selected = selectedIndex >= 0
+                ? this.Items[selectedIndex]
+                : this.Items.First();
 
             // initialize UI
-            this.Origin = new Vector2(x, y);
+            this.bounds.X = x;
+            this.bounds.Y = y;
             this.ReinitializeComponents();
         }
 
@@ -98,11 +119,12 @@ namespace Pathoschild.Stardew.Common.UI
         /// <returns>Returns whether an item was selected.</returns>
         public bool TrySelect(int x, int y, out TItem selected)
         {
-            foreach (ClickableComponent component in this.ItemComponents)
+            for (int i = 0; i < this.ItemComponents.Count; i++)
             {
+                var component = this.ItemComponents[i];
                 if (component.containsPoint(x, y))
                 {
-                    this.SelectedIndex = int.Parse(component.name);
+                    this.Selected = this.Items[i];
                     selected = this.SelectedItem;
                     return true;
                 }
@@ -112,18 +134,37 @@ namespace Pathoschild.Stardew.Common.UI
             return false;
         }
 
+        /// <summary>Select an item in the list matching the given value.</summary>
+        /// <param name="value">The value to search.</param>
+        /// <returns>Returns whether an item was selected.</returns>
+        public bool TrySelect(TItem value)
+        {
+            var entry = this.Items.FirstOrDefault(p =>
+                (p.Value == null && value == null)
+                || p.Value?.Equals(value) == true
+            );
+
+            if (entry == null)
+                return false;
+
+            this.Selected = entry;
+            return true;
+        }
+
         /// <summary>Render the UI.</summary>
         /// <param name="sprites">The sprites to render.</param>
         /// <param name="opacity">The opacity at which to draw.</param>
         public void Draw(SpriteBatch sprites, float opacity = 1)
         {
             // draw dropdown items
-            foreach (ClickableComponent component in this.ItemComponents)
+            for (int i = 0; i < this.ItemComponents.Count; i++)
             {
+                var component = this.ItemComponents[i];
+
                 // draw background
                 if (component.containsPoint(Game1.getMouseX(), Game1.getMouseY()))
                     sprites.Draw(CommonSprites.DropDown.Sheet, component.bounds, CommonSprites.DropDown.HoverBackground, Color.White * opacity);
-                else if (component.name.Equals(this.SelectedIndex.ToString()))
+                else if (i == this.Selected.Index)
                     sprites.Draw(CommonSprites.DropDown.Sheet, component.bounds, CommonSprites.DropDown.ActiveBackground, Color.White * opacity);
                 else
                     sprites.Draw(CommonSprites.DropDown.Sheet, component.bounds, CommonSprites.DropDown.InactiveBackground, Color.White * opacity);
@@ -141,29 +182,23 @@ namespace Pathoschild.Stardew.Common.UI
                 sprites.Draw(CommonSprites.Icons.Sheet, new Vector2(this.bounds.X - CommonSprites.Icons.UpArrow.Width, this.bounds.Y + this.bounds.Height - CommonSprites.Icons.DownArrow.Height), CommonSprites.Icons.DownArrow, Color.White * opacity);
         }
 
-
-        /*********
-        ** Private methods
-        *********/
         /// <summary>Recalculate dimensions and components for rendering.</summary>
-        private void ReinitializeComponents()
+        public void ReinitializeComponents()
         {
             // get item size
             int minItemWidth = Game1.tileSize * 2;
-            int itemWidth = Math.Max((int)this.Items.Max(p => this.Font.MeasureString(p.Name).X), minItemWidth) + DropdownList<TItem>.DropdownPadding * 2;
-            int itemHeight = this.FontHeight;
+            this.MaxLabelWidth = Math.Max((int)this.Items.Max(p => this.Font.MeasureString(p.Name).X), minItemWidth) + DropdownList<TItem>.DropdownPadding * 2;
+            int itemHeight = this.MaxLabelHeight;
 
             // get pagination
             int itemCount = this.Items.Length;
-            this.MaxItems = Math.Min((Game1.viewport.Height - (int)this.Origin.Y) / itemHeight, itemCount);
+            this.MaxItems = Math.Min((Game1.viewport.Height - (int)this.bounds.Y) / itemHeight, itemCount);
             this.MaxFirstVisibleIndex = this.Items.Length - this.MaxItems;
             this.FirstVisibleIndex = this.GetValidFirstItem(this.FirstVisibleIndex, this.MaxFirstVisibleIndex);
 
             // get dropdown size
-            this.bounds.Width = itemWidth;
+            this.bounds.Width = this.MaxLabelWidth;
             this.bounds.Height = itemHeight * this.MaxItems;
-            this.bounds.X = (int)this.Origin.X;
-            this.bounds.Y = (int)this.Origin.Y;
 
             // generate components
             this.ItemComponents.Clear();
@@ -171,11 +206,42 @@ namespace Pathoschild.Stardew.Common.UI
             int y = this.bounds.Y;
             for (int i = this.FirstVisibleIndex; i < this.MaxItems; i++)
             {
-                this.ItemComponents.Add(new ClickableComponent(new Rectangle(x, y, itemWidth, itemHeight), i.ToString()));
-                y += this.FontHeight;
+                this.ItemComponents.Add(new ClickableComponent(new Rectangle(x, y, this.MaxLabelWidth, itemHeight), i.ToString()));
+                y += this.MaxLabelHeight;
+            }
+
+            // update controller flow
+            this.ReinitializeControllerFlow();
+        }
+
+        /// <summary>Set the fields to support controller snapping.</summary>
+        public void ReinitializeControllerFlow()
+        {
+            int initialId = 1_100_000;
+            for (int last = this.ItemComponents.Count - 1, i = last; i >= 0; i--)
+            {
+                var component = this.ItemComponents[i];
+
+                component.myID = initialId + i;
+                component.upNeighborID = i != 0
+                    ? initialId + i - 1
+                    : -99999;
+                component.downNeighborID = i != last
+                    ? initialId + i + 1
+                    : -1;
             }
         }
 
+        /// <summary>Get the nested components for controller snapping.</summary>
+        public IEnumerable<ClickableComponent> GetChildComponents()
+        {
+            return this.ItemComponents;
+        }
+
+
+        /*********
+        ** Private methods
+        *********/
         /// <summary>Scroll the dropdown by the specified amount.</summary>
         /// <param name="amount">The number of items to scroll.</param>
         private void Scroll(int amount)